TST: Add test
